Dissociate clones by default; clone by arbitrary URL or by NAME; simplify add/remove
Using --reference by itself results in small but unsafe clones, and the user may
not realize that. If the cache directory is removed or the git folder is moved
off of the machine, the git repo will break. Move this functionality to a new
'refclone' command and instead by default make clones that are larger, but have
no dependencies upon the cache itself.

Allow for the use of arbitrary URLs when cloning that do not match those in the
cache. This is particularly useful for forks with a large base such as the linux
kernel. Keep support for names as an easy reference.

Remove the 'remote' command set and merge its commands into the primary set.
The use of the 'remote' keyword doesn't make sense in this context and just
makes for more typing.